+++ bpf/libbpf/kprobe.c
@@ -12,18 +12,27 @@
 #if defined(KPROBE)
 #include "kprobe.skel.h"
 #define struct_bpf	kprobe_bpf
+#define _bpf__open	kprobe_bpf__open
+#define _bpf__open_opts	kprobe_bpf__open_opts
+#define _bpf__load	kprobe_bpf__load
 #define _bpf__open_and_load	kprobe_bpf__open_and_load
 #define _bpf__attach	kprobe_bpf__attach
 #define _bpf__destroy	kprobe_bpf__destroy
 #elif defined(FENTRY)
 #include "fentry.skel.h"
 #define struct_bpf	fentry_bpf
+#define _bpf__open	fentry_bpf__open
+#define _bpf__open_opts	fentry_bpf__open_opts
+#define _bpf__load	fentry_bpf__load
 #define _bpf__open_and_load	fentry_bpf__open_and_load
 #define _bpf__attach	fentry_bpf__attach
 #define _bpf__destroy	fentry_bpf__destroy
 #elif defined(FMOD_RET)
 #include "fmod_ret.skel.h"
 #define struct_bpf	fmod_ret_bpf
+#define _bpf__open	fmod_ret_bpf__open
+#define _bpf__open_opts	fmod_ret_bpf__open_opts
+#define _bpf__load	fmod_ret_bpf__load
 #define _bpf__open_and_load	fmod_ret_bpf__open_and_load
 #define _bpf__attach	fmod_ret_bpf__attach
 #define _bpf__destroy	fmod_ret_bpf__destroy
@@ -45,11 +54,8 @@ int main(int argc, char **argv)
 
 	libbpf_set_print(libbpf_print_fn);
 
-	skel = _bpf__open_and_load();
-	if (!skel) {
-		fprintf(stderr, "Failed to open BPF skeleton\n");
-		return 1;
-	}
+	BPF__OPEN_AND_LOAD(skel, _bpf__open_and_load, _bpf__open_opts,
+			_bpf__load, _bpf__destroy);
 
 	err = _bpf__attach(skel);
 	if (err) {