+++ bpf/libbpf/sk_skb.c
@@ -11,18 +11,19 @@
 #include <sys/resource.h>
 #include <sys/socket.h>
 #include <unistd.h>
+#include "libbpf_wrapper.h"
 #include "sk_skb.skel.h"
 
 int main(int argc, char *argv[])
 {
+	int err;
 	struct sk_skb_bpf *skel;
 	int verdict;
 
-	skel = sk_skb_bpf__open_and_load();
-	if (!skel) {
-		fprintf(stderr, "open and load error\n");
-		return -1;
-	}
+	BPF__OPEN_AND_LOAD(skel, sk_skb_bpf__open_and_load,
+			sk_skb_bpf__open_opts,
+			sk_skb_bpf__load,
+			sk_skb_bpf__destroy);
 
 	verdict = bpf_program__fd(skel->progs.prog_skb_drop);