+++ bpf/libbpf/sockops.c
@@ -92,11 +92,10 @@ int main(int argc, char **argv)
 		return 1;
 	}
 
-	skel = sockops_bpf__open_and_load();
-	if (!skel) {
-		fprintf(stderr, "Failed to open BPF skeleton\n");
-		return 1;
-	}
+	BPF__OPEN_AND_LOAD(skel, sockops_bpf__open_and_load,
+			sockops_bpf__open_opts,
+			sockops_bpf__load,
+			sockops_bpf__destroy);
 
 	err = sockops_bpf__attach(skel);
 	if (err) {